+++ components/21-atoms/logo/logo.config.js
@@ -10,7 +10,7 @@ module.exports = {
       name: 'subsite',
       context: {
         type: 'subsite',
-        site_name: 'ICT supplier Digipolis Gent'
+        site_name: 'ICT supplier District09'
       }
     }
   ]